Add functions for restoring scroll pos when switching buffers
By default vim will bring your current line to the center of the
screen when switching between buffers. I find this very annoying b/c it
breaks my spacial awareness of where I was when I was last looking at
the file. It breaks my workflow of “look at this”, “look at that”, “look
back at this”. This forces vim to keep the current line AND scrolls
position when switching between buffers